Add files via upload
One of the most fundamental operations in computer vision is to convert a RGB 3 channel image to a grayscale single channel image. Here I have assumed a linear RGB colorspace of the input image. If it is a non-linear RGB,then first gamma uncompression has to be applied, then use the linear conversion formula and again apply gamma compression to get the desired result.